Reorganizing, trying new things.
I'm going to try to have offline urls scanned three times before
they're ignored.
